Recursively copy files.

+++ install.py
@@ -1,4 +1,8 @@
 #!/usr/bin/python
+#
+# Installs all the files and any template directories in 
+# the same directory as this script to the home directory. 
+# 
 
 import getopt
 import glob
@@ -7,11 +11,14 @@ import shutil
 from subprocess import Popen, PIPE
 import sys
 
+EXCLUDES = ['.git', 'README', 'install.py']
+DIFF = 'diff -u'
+COLORDIFF = 'colordiff'
+
 # thanks, stackoverflow!
 def is_on_path(program):
     def can_execute(fpath):
         return os.path.exists(fpath) and os.access(fpath, os.X_OK)
-
     fpath, fname = os.path.split(program)
     if fpath:
         if can_execute(program):
@@ -30,43 +37,78 @@ def diff(path1, path2):
         sys.exit(1)
     output = None
     if not colordiff_path:
-        p = Popen(["diff", "-u", path1, path2], stdin=PIPE, stdout=PIPE)
+        p = Popen(DIFF.split() + [path1, path2], stdin=PIPE, stdout=PIPE)
         p.stdin.close()
         o = p.stdout.read()
         if p.wait() != 0:
             output = o
     else:
-        p1 = Popen(["diff", "-u", path1, path2], stdin=PIPE, stdout=PIPE)
-        p2 = Popen(["colordiff"], stdin=p1.stdout, stdout=PIPE)
+        p1 = Popen(DIFF.split() + [path1, path2], stdin=PIPE, stdout=PIPE)
+        p2 = Popen(COLORDIFF.split(), stdin=p1.stdout, stdout=PIPE)
         o = p2.communicate()[0]
         if p1.wait() != 0:
             output  = o
     return output
 
-def install_dotfile(srcpath, force=False, pretend=False):
-    def install(src, dest):
-        if not pretend:
-            shutil.copyfile(src, dest)
-        print("Installed %s" % tildepath)
-
+def install_entry(srcpath, force=False, pretend=False):
+    entry_type = None
+    if os.path.isfile(srcpath):
+        entry_type = 'File'
+    elif os.path.isdir(srcpath):
+        entry_type = 'Directory'
     installed = False
     different = False
     tildepath = "~/%s" % srcpath
     destpath = os.path.expanduser(tildepath)
+
+    def install(src, dest):
+        skipped = False
+        if os.path.isfile(src):
+            if os.path.exists(dest) and not os.path.isfile(dest):
+                skipped = True
+            else:
+                if not pretend:
+                    shutil.copyfile(src, dest)
+                print("Installed File %s" % tildepath)
+        elif os.path.isdir(src):
+            if os.path.exists(dest) and not os.path.isdir(dest):
+                skipped = True
+            else:
+                if not pretend and not os.path.exists(dest):
+                    os.makedirs(dest, mode=0775)
+                print("Installed Directory %s" % tildepath)
+        if skipped:
+            print("Skipped %s %s, destination exists and is not a %s" % (entry_type, tildepath, entry_type))
+
     if os.path.exists(destpath):
         if force:
             install(srcpath, destpath)
             installed = True
         else:
             print("Skipped %s" % tildepath)
-            diff_output = diff(destpath, srcpath)
-            if diff_output:
-                different = True
-                print(diff_output)
+            if os.path.isfile(srcpath) and os.path.isfile(destpath):
+                diff_output = diff(destpath, srcpath)
+                if diff_output:
+                    different = True
+                    print(diff_output)
     else:
         install(srcpath, destpath)
         installed = True
     return (installed, different)
+
+def strip_cwd(path):
+    path = path.replace(os.getcwd(), '')
+    if path[0] == '/': path = path[1:]
+    return path
+
+def recursive_entries(basepath, excludes=EXCLUDES):
+    for item in [os.path.join(basepath, x)
+                 for x in os.listdir(basepath)
+                 if not os.path.basename(x) in excludes]:
+        yield strip_cwd(item)
+        if os.path.isdir(item):
+            for child_item in recursive_entries(item):
+                yield strip_cwd(child_item)
 
 if __name__ == "__main__":
     force = False
@@ -80,9 +122,10 @@ if __name__ == "__main__":
         if o == "-f": force = True
         elif o == "-p": pretend = True
 
-    results = [install_dotfile(fn, force=force, pretend=pretend)
-               for fn in glob.glob('.*')
-               if fn != '.git']
+    script_dir = os.path.abspath(os.path.dirname(sys.argv[0]))
+
+    results = [install_entry(fn, force=force, pretend=pretend)
+              for fn in recursive_entries(script_dir)]
 
     all_installed = all(x[0] for x in results) 
     any_different = any(x[1] for x in results)
